tests/wpt/tests/navigation-api/navigate-event/signal-abort-window-stop-after-intercept.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    let start_url = location.href;
    let abort_signal;
    let onabort_called = false;
    let navigateErrorException;
    navigation.onnavigateerror = t.step_func(e => {
      assert_equals(e.constructor, ErrorEvent);
      navigateErrorException = e.error;
      assert_equals(e.filename, start_url);
      assert_greater_than(e.lineno, 0);
      assert_greater_than(e.colno, 0);
    });
    navigation.onnavigatesuccess = t.unreached_func("onnavigatesuccess");
    navigation.onnavigate = t.step_func(e => {
      abort_signal = e.signal;
      abort_signal.onabort = () => onabort_called = true;
      e.intercept({ handler: () => new Promise(resolve => t.step_timeout(resolve, 0)) });
    });
    let result = navigation.navigate("?1");
    window.stop();
    assert_true(abort_signal.aborted);
    assert_true(onabort_called);

    result.committed.then(() => {
      return promise_rejects_dom(t, 'AbortError', result.finished);
    }).then(() => {
      return result.finished.catch(e => assert_equals(e, navigateErrorException));
    }).then(() => {
      // Complete the test asynchronously to ensure that onnavigatesuccess
      // didn't fire on a microtask.
      t.step_timeout(t.step_func_done(() => {}), 5);
    });
  });
}, "window.stop() cancels the navigate event's intercept() and signals event.signal");
</script>
