LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/signal-abort-window-stop-after-intercept-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: expected (undefined) undefined but got (object) object "AbortError: Navigation aborted"

Harness Error (FAIL), message = Unhandled rejection: assert_equals: expected (undefined) undefined but got (object) object "AbortError: Navigation aborted"

TIMEOUT window.stop() cancels the navigate event's intercept() and signals event.signal Test timed out

Harness Error (FAIL), message = Unhandled rejection: assert_equals: expected (undefined) undefined but got (object) object "AbortError: Navigation aborted"

TIMEOUT window.stop() cancels the navigate event's intercept() and signals event.signal Test timed out
